requirements.txt
Django==1.5.5
Markdown==2.3.1
South==0.8.2
django-filter==0.7
django-suit==0.2.5
djangorestframework==2.3.8
ipdb==0.8
ipython==1.1.0
wsgiref==0.1.2
